+++ src/dipdup/datasources/metadata/datasource.py
@@ -10,8 +10,8 @@ from dipdup.datasources.metadata.enums import MetadataNetwork
 
 class MetadataDatasource(GraphQLDatasource):
     _default_http_config = HTTPConfig(
-        retry_count=3,
         retry_sleep=1,
+        retry_count=3,
         ratelimit_rate=10,
         ratelimit_period=1,
     )