+++ src/dipdup/dipdup.py
@@ -434,8 +434,11 @@ class DipDup:
             except KeyError:
                 await self._ctx.reindex(ReindexingReason.schema_modified)
 
-        # NOTE: Call even if Schema is present; there may be new tables
-        await generate_schema(conn, schema_name)
+        # NOTE: Call even if Schema is present to create new tables
+        await generate_schema(
+            conn,
+            schema_name,
+        )
         schema_hash = get_schema_hash(conn)
 
         if self._schema is None: